+++ include/io/JsonBuilder.hpp
@@ -48,10 +48,10 @@ public:
          double /*elapsed*/,
          int /*transform_size*/);
 
-    static std::vector<uint32_t> compactBits(
+    static std::vector<uint64_t> compactBits(
         const std::vector<uint64_t>& x,
         const std::vector<int>& digit_width,
-        uint32_t E);
+        uint64_t E);
 };
 
 } // namespace io
